+++ mangooio-core/src/main/java/io/mangoo/routing/handlers/RequestHandler.java
@@ -13,10 +13,12 @@ import io.mangoo.routing.bindings.Request;
 import io.mangoo.templating.TemplateContext;
 import io.mangoo.utils.JsonUtils;
 import io.mangoo.utils.RequestUtils;
+import io.mangoo.utils.internal.MangooUtils;
 import io.mangoo.utils.internal.Trace;
 import io.undertow.server.HttpHandler;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.server.handlers.Cookie;
+import jakarta.validation.ConstraintViolation;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.logging.log4j.util.Strings;
@@ -30,6 +32,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 
 public class RequestHandler implements HttpHandler {
 
@@ -132,6 +135,16 @@ public class RequestHandler implements HttpHandler {
             invokedResponse = (Response) attachment.getMethod().invoke(attachment.getControllerInstance());
         } else {
             final Object [] convertedParameters = getConvertedParameters(exchange);
+            Set<ConstraintViolation<Object>> violations =
+                    MangooUtils.validator().validateParameters(
+                            attachment.getControllerInstance(),
+                            attachment.getMethod(),
+                            convertedParameters);
+
+            if (!violations.isEmpty()) {
+                return Response.badRequest().bodyDefault().end();
+            }
+
             invokedResponse = (Response) attachment.getMethod().invoke(attachment.getControllerInstance(), convertedParameters);
         }